+++ test/e2e/specs/editor/collaboration/collaboration-title-reload.spec.ts
@@ -33,6 +33,7 @@ const test = base.extend< Fixtures >( {
 		await setCollaboration( requestUtils, true );
 		await use( utils );
 		await utils.teardown();
+		await setCollaboration( requestUtils, false );
 	},
 	collaboratorUser: async (
 		{ collaborationUtils, requestUtils },